Fix deployment script for dapps.
